pessoa_fisica_creator_controller with tests

+++ src/models/sqlite/interfaces/client_interface.py
@@ -19,10 +19,10 @@ class ClientInterface(ABC):
     pass
 
   @abstractmethod
-  def deposit(self, id: int, deposit_value: int) -> None:
+  def deposit(self, id: int, deposit_value: float) -> None:
     pass
 
   @abstractmethod
-  def withdraw(self, id: int, withdraw_value: int) -> None:
+  def withdraw(self, id: int, withdraw_value: float) -> None:
     pass
