+++ src/models/sqlite/repositories/pessoa_juridica_repository.py
@@ -32,7 +32,7 @@ class PessoaJuridicaRepository(ClientInterface):
       except NoResultFound:
         return []
 
-  def insert_client(self, faturamento: int, idade: int, nome_fantasia: str, celular: str, email_corporativo: str, categoria: str, saldo: int = 0) -> None:
+  def insert_client(self, faturamento: int, idade: int, nome_fantasia: str, celular: str, email_corporativo: str, categoria: str, saldo: float = 0) -> None:
     with self.__db_connection as database:
       try:
         new_client_data = PessoaJuridicaTable(
@@ -57,7 +57,7 @@ class PessoaJuridicaRepository(ClientInterface):
     except Exception:
       return None
 
-  def deposit(self, id: int, deposit_value: int) -> None:
+  def deposit(self, id: int, deposit_value: float) -> None:
     with self.__db_connection as database:
       try:
         client = (
@@ -72,7 +72,7 @@ class PessoaJuridicaRepository(ClientInterface):
         database.session.rollback()
         raise exception
 
-  def withdraw(self, id: int, withdraw_value: int) -> None:
+  def withdraw(self, id: int, withdraw_value: float) -> None:
     with self.__db_connection as database:
       try:
         client = (
